+++ demo/source/main.d
@@ -60,7 +60,7 @@ class TestScene: Scene
     override void onFocusLoss() { }
     override void onFocusGain() { }
     override void onDropFile(string filename) { }
-    override void onUserEvent(int code) { }
+    override void onUserEvent(int code, void* payload) { }
     override void onQuit() { }
 }